requirements.txt
gpiozero
RPi.GPIO
tk
